requirements.txt
Flask==1.1.2
Keras==2.4.3
Flask_Cors==3.0.9
numpy==1.19.5
opencv_python==4.4.0.46
Werkzeug==1.0.1
tensorflow==2.4.0
pandas==1.1.5
Pillow==8.1.2
